implementation as a `trigger` instead of an `Extension` subclass.
